apps/backend/requirements.txt
requests
pandas
scikit-learn
joblib
fastapi
uvicorn
httpx
python-dotenv
beautifulsoup4
lxml
python-dateutil
playwright
pytest
tldextract
